# docs/swagger/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the FeatherWings daemon.",
        "title": "FeatherWings API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/servers": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "List servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.APIResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Create server",
                "parameters": [
                    {
                        "description": "Server configuration",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/installer.ServerDetails"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Get server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Delete server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/servers/{server}/files/chmod": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Server Files"
                ],
                "summary": "Change file permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chmod request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.ServerChmodRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/files/compress": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server Files"
                ],
                "summary": "Compress files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Archive request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.ServerArchiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/filesystem.Stat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/servers/{server}/files/pull/{download}": {
            "delete": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Server Files"
                ],
                "summary": "Cancel remote download",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Download identifier",
                        "name": "download",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "server.APIResponse": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/server.Configuration"
                },
                "is_suspended": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string"
                },
                "utilization": {
                    "$ref": "#/definitions/server.ResourceUsage"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "NodeToken": {
            "description": "Supply the node's bearer token from `config.yml` using the `Authorization: Bearer \u003ctoken\u003e` header.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerJWT": {
            "description": "Signed JWTs issued by the Panel for server-scoped operations (uploads, downloads, websockets). Pass in the `token` query parameter.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}